backend/requirements.txt
fastapi
uvicorn
httpx
pydantic
python-dotenv
tenacity
cachetools
google-generativeai
